requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.4
colorama==0.4.6
Django==4.2.1
django-jazzmin==2.6.0
et-xmlfile==1.1.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pymemcache==4.0.0
pywin32==306
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
xlwings==0.30.7
pygsheets
load_dotenv